% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{cache_compose}
\alias{cache_compose}
\title{Compose any number of cache objects into a new, layered cache object}
\usage{
cache_compose(...)
}
\arguments{
\item{...}{Cache objects to compose into a new, layered cache object.}
}
\description{
Compose any number of cache objects into a new, layered cache object
}
