% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache-layered.R
\name{cache_layered}
\alias{cache_layered}
\title{Compose any number of cache objects into a new, layered cache object}
\usage{
cache_layered(..., logfile = NULL)
}
\arguments{
\item{...}{Cache objects to compose into a new, layered cache object.}

\item{logfile}{An optional filename or connection object to where logging
information will be written. To log to the console, use \code{stderr()} or
\code{stdout()}.}
}
\description{
Compose any number of cache objects into a new, layered cache object
}
